Add snippet for inserting a code block.
Tab trigger: <kbd>`⇥</kbd> (that is, a single backtick and then pressing the tab key). Note that typing a single backtick will insert two backticks (with the cursor in between), because the Markdown bundle defines an appropriate smart typing pair. This is why the snippet inserts only two backticks at the end of the code block: Together with the backtick already inserted by the smart typing pair this gives three backticks at the end.

The snippet shows a completion list for specifying the language of the code block. The list of supported languages was extracted from the bundle’s grammar. If at some point in the future more languages are added to the grammar, these languages should be added to the completion list as well.